requirement.txt
flask
google-generativeai
gunicorn
pysqlite3
datetime
wikipedia
threading